Add new methods to Server, modify start behavour
isBound() for seeing if the server has initialised
stop() for cleanly disconnecting clients and cleaning up in one shot

Initialisation behaviour was modified so that it can be
repeated. This allows for (eg) test cases to JUST initialise without starting
the client connection loop, which contains blocking method calls. It also
keeps the ability for someone to call run() on Server and have initialisation
happen implicitly.