Major refactor for networking. No more hacky indices
Prior to this patch, the index of the Event in the enum's array
of values was used to communicate actions between client and
server. This was likely undefined behaviour and hackish at best.
I'm not convinced at all that there is any guarantee that each
end will produce the same array of enum members, all in the same
order.
This patch takes advantage of the Object{In,Out}putStream objects
to ensure this doesn't happen, by serialising and transferring
the objects much lik the Zone is with game state updates.