Consolidate Threads,  simplify disconnection
The downstream threads for keeping the client updated with the game
status were an afterthought or add-on and as such were not very well
integrated with the code. This patch merges both the upstream and
downstream worker threads (Send and Recv) into one array.

Additionally, the worker threads are no longer notified of a disconnection.
This behaviour was overly-cautious and it is perfectly legitimate to
simply let them error-out when the socket is closed.