Keep Object{Input,Output}Stream objects consistent
Previous behaviour was to create an Object{Input,Output}Stream
on the fly as needed. This was a naïve approach that ignored
the fact that the streams must be balanced on the client and
server side. This meant that a Client with stream ID A could talk to
the server with stream ID A, but if the server had a stream with ID B
then it would also receive these (apparently) invalid messages
since they weren't send from a matching stream. This would error-out.

This patch fixes this behaviour.