Use the player's facing direction as their new direction
This might be a perference thing, I'm not sure. But
at least while our graphics are too simple to display player
facing direction, strafing is a better idea.

It was very confusing to be switching between two cells when
repeatedly moving backwards.

This might want to be reverted when our graphics will make it
more clear what is going on.