Fix NullPointerException on certain packet errors
ServerRecvThread used to have a reference to the entire socket
passed, which is was constructed with. This was removed in favour
of passing only an ObjectOutputStream. However, the socket field
was never removed and reference was still made to it within the
class's code. This field remained uninitialised throughout the
lifetime of the thread, so any reference to it (specifically in
the error messages on invalid/incorrect types for packets being
sent.

This patch removes references to the socket field and indeed
the field itself.