Do not send game objects from client to server
Sending objects with important game state from the client to
the server and subsequently acting on these objects is not
the behaviour we normally want. This will result in this
object being local to the ServerRecvThread being used and
potentially modified, never changing the objects in the Zone
that is being broadcast to clients.

This resulted in some behaviour that looked very interesting
and was confusing until it clicked.